+++ src/application/controller/user.controller.js
@@ -1,5 +1,15 @@
-import userService from './../services/user.services.js'
+import userService from './../services/user.services.js';
+
 const searchUsers = (request, response) => {
+    if (request.auth.role !== 1) {
+        response.status(403)
+            .send({
+                error: 'Không có quyền truy cập.'
+            })
+
+        return;
+    }
+
     const { name, page, limit } = request.query;
 
     userService.searchUsers({ name: name, page: page, limit: limit }, (error, result) => {
@@ -12,15 +22,24 @@ const searchUsers = (request, response) => {
             response.send(result);
         }
     })
-
-
 }
+
 const addUser = (request, response) => {
+    if (request.auth.role !== 1) {
+        response.status(403)
+            .send({
+                error: 'Không có quyền truy cập.'
+            })
+
+        return;
+    }
+
     const requestBody = request.body;
     const avatar = request.file;
 
     userService.addUser({
         ...requestBody,
+        authId: request.auth.user_id,
         avatar: avatar
     }, (error, result) => {
         if (error) {
@@ -32,9 +51,19 @@ const addUser = (request, response) => {
             response.status(201)
                 .send();
         }
-    })
+    });
 }
+
 const getDetailUser = (request, response) => {
+    if (request.auth.role !== 1) {
+        response.status(403)
+            .send({
+                error: 'Không có quyền truy cập.'
+            })
+
+        return;
+    }
+
     const { id } = request.params;
     userService.getDetailUser(id, (error, result) => {
         if (error) {
@@ -46,24 +75,49 @@ const getDetailUser = (request, response) => {
             response.send(result);
         }
     })
-
 }
+
 const updateUser = (request, response) => {
-    const { id } = request.params;
-    const requestBody = request.body;
-    userService.updateUser(id, requestBody, (error, result) => {
-        if (error) {
-            response.status(500).send({
-                error: error,
+    if (request.auth.role !== 1) {
+        response.status(403)
+            .send({
+                error: 'Không có quyền truy cập.'
             })
-        } else {
-            response.status(204).send()
-        }
-    })
 
+        return;
+    }
+
+    const userId = request.params.id;
+    const requestBody = request.body;
+    const avatar = request.file;
+
+    userService.updateUser(userId, {
+        ...requestBody,
+        avatar: avatar
+    }, (error, result) => {
+        if (error) {
+            response.status(500)
+                .send({
+                    error: error,
+                });
+        } else {
+            response.status(200)
+                .send();
+        }
+    });
 
 }
+
 const deleteUser = (request, response) => {
+    if (request.auth.role !== 1) {
+        response.status(403)
+            .send({
+                error: 'Không có quyền truy cập.'
+            })
+
+        return;
+    }
+
     const { id } = request.params;
 
     userService.deleteUser(id, (error, result) => {
@@ -76,12 +130,12 @@ const deleteUser = (request, response) => {
             response.status(204).send();
         }
     })
-
 }
+
 export default {
     searchUsers,
     addUser,
     getDetailUser,
     updateUser,
-    deleteUser
+    deleteUser,
 }
